@@ -93,49 +93,48 @@
 		"@types/request": "2.48.3",
 		"@types/request-promise-native": "1.0.17",
 		"@types/request-stats": "3.0.0",
-		"@types/rimraf": "2.0.2",
+		"@types/rimraf": "2.0.3",
 		"@types/seedrandom": "2.4.28",
-		"@types/sharp": "0.22.3",
+		"@types/sharp": "0.23.0",
 		"@types/showdown": "1.9.3",
 		"@types/speakeasy": "2.0.5",
-		"@types/systeminformation": "3.23.1",
+		"@types/systeminformation": "3.54.1",
 		"@types/tinycolor2": "1.4.2",
 		"@types/tmp": "0.1.0",
-		"@types/uuid": "3.4.5",
 		"@types/vue": "2.0.0",
 		"@types/web-push": "3.3.0",
-		"@types/webpack": "4.39.3",
+		"@types/webpack": "4.41.0",
 		"@types/webpack-stream": "3.2.10",
-		"@types/websocket": "0.0.40",
+		"@types/websocket": "1.0.0",
 		"@types/ws": "6.0.3",
-		"@typescript-eslint/parser": "2.3.3",
+		"@typescript-eslint/parser": "2.8.0",
 		"agentkeepalive": "4.1.0",
 		"animejs": "3.1.0",
-		"apexcharts": "3.10.0",
+		"apexcharts": "3.10.1",
 		"autobind-decorator": "2.4.0",
 		"autosize": "4.0.2",
 		"autwh": "0.1.0",
-		"aws-sdk": "2.548.0",
+		"aws-sdk": "2.578.0",
 		"bcryptjs": "2.4.3",
 		"bootstrap": "4.3.1",
-		"bootstrap-vue": "2.0.4",
-		"bull": "3.11.0",
-		"cafy": "15.1.1",
+		"bootstrap-vue": "2.1.0",
+		"bull": "3.12.1",
+		"cafy": "15.2.0",
 		"cbor": "5.0.1",
 		"chai": "4.2.0",
-		"chalk": "2.4.2",
-		"cli-highlight": "2.1.1",
-		"commander": "3.0.2",
+		"chalk": "3.0.0",
+		"cli-highlight": "2.1.4",
+		"commander": "4.0.1",
 		"content-disposition": "0.5.3",
 		"crc-32": "1.2.0",
 		"css-loader": "3.2.0",
 		"cssnano": "4.1.10",
 		"dateformat": "3.0.3",
-		"deep-equal": "1.1.0",
+		"deep-equal": "1.1.1",
 		"diskusage": "1.1.3",
 		"double-ended-queue": "2.1.0-0",
-		"eslint": "6.5.1",
-		"eslint-plugin-vue": "5.2.3",
+		"eslint": "6.7.0",
+		"eslint-plugin-vue": "6.0.1",
 		"eventemitter3": "4.0.0",
 		"feed": "4.0.0",
 		"file-type": "12.3.0",
@@ -230,54 +228,54 @@
 		"stylus-loader": "3.0.2",
 		"summaly": "2.3.1",
 		"syslog-pro": "1.0.0",
-		"systeminformation": "4.14.11",
+		"systeminformation": "4.15.3",
 		"syuilo-password-strength": "0.0.1",
-		"terser-webpack-plugin": "2.1.3",
+		"terser-webpack-plugin": "2.2.1",
 		"textarea-caret": "3.1.0",
-		"three": "0.109.0",
+		"three": "0.110.0",
 		"tinycolor2": "1.4.1",
 		"tmp": "0.1.0",
-		"ts-loader": "6.2.0",
-		"ts-node": "8.4.1",
-		"tslint": "5.20.0",
+		"ts-loader": "6.2.1",
+		"ts-node": "8.5.2",
+		"tslint": "5.20.1",
 		"tslint-sonarts": "1.9.0",
-		"typeorm": "0.2.19",
-		"typescript": "3.6.4",
+		"typeorm": "0.2.20",
+		"typescript": "3.7.2",
 		"uglify-es": "3.3.9",
 		"ulid": "2.3.0",
-		"url-loader": "2.2.0",
+		"url-loader": "2.3.0",
 		"uuid": "3.3.3",
 		"v-animate-css": "0.0.3",
 		"v-debounce": "0.1.2",
 		"vue": "2.6.10",
 		"vue-color": "2.7.0",
 		"vue-content-loading": "1.6.0",
-		"vue-cropperjs": "4.0.0",
-		"vue-i18n": "8.14.1",
+		"vue-cropperjs": "4.0.1",
+		"vue-i18n": "8.15.0",
 		"vue-js-modal": "1.3.31",
 		"vue-json-pretty": "1.6.2",
-		"vue-loader": "15.7.1",
+		"vue-loader": "15.7.2",
 		"vue-marquee-text-component": "1.1.1",
 		"vue-prism-component": "1.1.1",
 		"vue-router": "3.1.3",
 		"vue-sequential-entrance": "1.1.3",
 		"vue-style-loader": "4.1.2",
-		"vue-svg-inline-loader": "1.3.3",
+		"vue-svg-inline-loader": "1.4.3",
 		"vue-template-compiler": "2.6.10",
 		"vue-touch": "2.0.0-beta.4",
 		"vuedraggable": "2.23.2",
 		"vuewordcloud": "18.7.11",
-		"vuex": "3.1.1",
-		"vuex-persistedstate": "2.5.4",
-		"web-push": "3.4.0",
-		"webpack": "4.41.1",
-		"webpack-cli": "3.3.9",
+		"vuex": "3.1.2",
+		"vuex-persistedstate": "2.7.0",
+		"web-push": "3.4.1",
+		"webpack": "4.41.2",
+		"webpack-cli": "3.3.10",
 		"websocket": "1.0.30",
-		"ws": "7.1.2",
+		"ws": "7.2.0",
 		"xev": "2.0.1"
 	},
 	"devDependencies": {
-		"@types/fluent-ffmpeg": "2.1.10",
-		"npm-run-all": "4.1.5"
+		"npm-run-all": "4.1.5",
+		"@types/fluent-ffmpeg": "2.1.11"
 	}
 }
