@@ -11,8 +11,7 @@ const cssnano = require('gulp-cssnano');
 const stylus = require('gulp-stylus');
 import * as uglifyComposer from 'gulp-uglify/composer';
 import * as rimraf from 'rimraf';
-import chalk from 'chalk';
-const imagemin = require('gulp-imagemin');
+import * as chalk from 'chalk';
 import * as rename from 'gulp-rename';
 import * as mocha from 'gulp-mocha';
 import * as replace from 'gulp-replace';
@@ -97,7 +96,7 @@ gulp.task('cleanall', gulp.parallel('clean', cb =>
 ));
 
 gulp.task('build:client:script', () => {
-	const client = require('./built/client/meta.json');
+	const client = require('./built/meta.json');
 	return gulp.src(['./src/client/app/boot.js', './src/client/app/safe.js'])
 		.pipe(replace('VERSION', JSON.stringify(client.version)))
 		.pipe(replace('ENV', JSON.stringify(env)))
@@ -122,7 +121,6 @@ gulp.task('copy:client', () =>
 			'./src/client/assets/**/*',
 			'./src/client/app/*/assets/**/*'
 		])
-			.pipe(isProduction ? (imagemin as any)() : gutil.noop())
 			.pipe(rename(path => {
 				path.dirname = path.dirname!.replace('assets', '.');
 			}))
