@@ -1,5 +1,5 @@
 import * as cluster from 'cluster';
-import chalk from 'chalk';
+import * as chalk from 'chalk';
 import Xev from 'xev';
 
 import Logger from '../services/logger';
