@@ -4,7 +4,7 @@
 	<div class="popover" :class="{ isMobile: $root.isMobile }" ref="popover">
 		<p v-if="!$root.isMobile">{{ title }}</p>
 		<div class="buttons" ref="buttons" :class="{ showFocus }">
-			<button v-for="(reaction, i) in $store.state.settings.reactions" :key="reaction" @click="react(reaction)" @mouseover="onMouseover" @mouseout="onMouseout" :tabindex="i + 1" :title="/^[a-z]+$/.test(reaction) ? $t('@.reactions.' + reaction) : reaction" v-particle><mk-reaction-icon :reaction="reaction"/></button>
+			<button v-for="(reaction, i) in rs" :key="reaction" @click="react(reaction)" @mouseover="onMouseover" @mouseout="onMouseout" :tabindex="i + 1" :title="/^[a-z]+$/.test(reaction) ? $t('@.reactions.' + reaction) : reaction" v-particle><mk-reaction-icon :reaction="reaction"/></button>
 		</div>
 		<div v-if="enableEmojiReaction" class="text">
 			<input v-model="text" :placeholder="$t('input-reaction-placeholder')" @keyup.enter="reactText" @input="tryReactText" v-autocomplete="{ model: 'text' }">
@@ -22,16 +22,11 @@ import { emojiRegex } from '../../../../../misc/emoji-regex';
 export default Vue.extend({
 	i18n: i18n('common/views/components/reaction-picker.vue'),
 	props: {
-		note: {
-			type: Object,
-			required: true
-		},
-
 		source: {
 			required: true
 		},
 
-		cb: {
+		reactions: {
 			required: false
 		},
 
@@ -50,6 +45,7 @@ export default Vue.extend({
 
 	data() {
 		return {
+			rs: this.reactions || this.$store.state.settings.reactions,
 			title: this.$t('choose-reaction'),
 			text: null,
 			enableEmojiReaction: false,
@@ -134,14 +130,7 @@ export default Vue.extend({
 
 	methods: {
 		react(reaction) {
-			this.$root.api('notes/reactions/create', {
-				noteId: this.note.id,
-				reaction: reaction
-			}).then(() => {
-				if (this.cb) this.cb();
-				this.$emit('closed');
-				this.destroyDom();
-			});
+			this.$emit('chosen', reaction);
 		},
 
 		reactText() {
@@ -221,7 +210,6 @@ export default Vue.extend({
 		width 100%
 		height 100%
 		background var(--modalBackdrop)
-		backdrop-filter blur(4px)		
 		opacity 0
 
 	> .popover
