@@ -3,8 +3,10 @@
 	<x-pie class="pie" :value="usage"/>
 	<div>
 		<p><fa icon="microchip"/>CPU</p>
-		<p>{{ meta.cpu.cores }} Logical cores</p>
-		<p>{{ meta.cpu.model }}</p>
+		<template v-if="meta.cpu">
+			<p>{{ meta.cpu.cores }} Logical cores</p>
+			<p>{{ meta.cpu.model }}</p>
+		</template>
 	</div>
 </div>
 </template>
