@@ -26,7 +26,7 @@
 					<mfm v-if="user.description" :text="user.description" :is-note="false" :author="user" :i="$store.state.i" :custom-emojis="user.emojis" :key="user.id"/>
 					<x-integrations :user="user" style="margin:20px 0;"/>
 				</div>
-				<div class="fields" v-if="user.fields">
+				<div class="fields" v-if="user.fields" :key="user.id">
 					<dl class="field" v-for="(field, i) in user.fields" :key="i">
 						<dt class="name">
 							<mfm :text="field.name" :plain="true" :custom-emojis="user.emojis"/>
