@@ -183,8 +183,8 @@ export function toHtml(tokens: MfmForest | null, mentionedRemoteUsers: IMentione
 					break;
 				default:
 					const remoteUserInfo = mentionedRemoteUsers.find(remoteUser => remoteUser.username === username && remoteUser.host === host);
-					a.href = remoteUserInfo ? remoteUserInfo.uri : `${config.url}/${acct}`;
-					a.className = 'mention';
+					a.href = remoteUserInfo ? (remoteUserInfo.url ? remoteUserInfo.url : remoteUserInfo.uri) : `${config.url}/${acct}`;
+					a.className = 'u-url mention';
 					break;
 			}
 			a.textContent = acct;
