@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 import * as request from 'request';
 import config from '../config';
-import chalk from 'chalk';
+import * as chalk from 'chalk';
 import Logger from '../services/logger';
 
 export async function downloadUrl(url: string, path: string) {
