@@ -3,8 +3,7 @@ import * as promiseLimit from 'promise-limit';
 import config from '../../../config';
 import Resolver from '../resolver';
 import { resolveImage } from './image';
-import { isCollectionOrOrderedCollection, isCollection, IPerson } from '../type';
-import { DriveFile } from '../../../models/entities/drive-file';
+import { isCollectionOrOrderedCollection, isCollection, IPerson, getApId } from '../type';
 import { fromHtml } from '../../../mfm/fromHtml';
 import { resolveNote, extractEmojis } from './note';
 import { registerOrFetchInstanceDoc } from '../../../services/register-or-fetch-instance-doc';
@@ -27,6 +26,7 @@ import { validActor } from '../../../remote/activitypub/type';
 import { getConnection } from 'typeorm';
 import { ensure } from '../../../prelude/ensure';
 import { toArray } from '../../../prelude/array';
+import { fetchNodeinfo } from '../../../services/fetch-nodeinfo';
 
 const logger = apLogger;
 
@@ -150,13 +150,13 @@ export async function createPerson(uri: string, resolver?: Resolver): Promise<Us
 				createdAt: new Date(),
 				lastFetchedAt: new Date(),
 				name: person.name,
-				isLocked: person.manuallyApprovesFollowers,
+				isLocked: !!person.manuallyApprovesFollowers,
 				username: person.preferredUsername,
-				usernameLower: person.preferredUsername.toLowerCase(),
+				usernameLower: person.preferredUsername!.toLowerCase(),
 				host,
 				inbox: person.inbox,
 				sharedInbox: person.sharedInbox || (person.endpoints ? person.endpoints.sharedInbox : undefined),
-				featured: person.featured,
+				featured: person.featured ? getApId(person.featured) : undefined,
 				uri: person.id,
 				tags,
 				isBot,
@@ -191,6 +191,7 @@ export async function createPerson(uri: string, resolver?: Resolver): Promise<Us
 	registerOrFetchInstanceDoc(host).then(i => {
 		Instances.increment({ id: i.id }, 'usersCount', 1);
 		instanceChart.newUser(i.host);
+		fetchNodeinfo(i);
 	});
 
 	usersChart.update(user!, true);
@@ -199,18 +200,18 @@ export async function createPerson(uri: string, resolver?: Resolver): Promise<Us
 	updateUsertags(user!, tags);
 
 	//#region アバターとヘッダー画像をフェッチ
-	const [avatar, banner] = (await Promise.all<DriveFile | null>([
+	const [avatar, banner] = await Promise.all([
 		person.icon,
 		person.image
 	].map(img =>
 		img == null
 			? Promise.resolve(null)
 			: resolveImage(user!, img).catch(() => null)
-	)));
+	));
 
 	const avatarId = avatar ? avatar.id : null;
 	const bannerId = banner ? banner.id : null;
-	const avatarUrl = avatar ? DriveFiles.getPublicUrl(avatar) : null;
+	const avatarUrl = avatar ? DriveFiles.getPublicUrl(avatar, true) : null;
 	const bannerUrl = banner ? DriveFiles.getPublicUrl(banner) : null;
 	const avatarColor = avatar && avatar.properties.avgColor ? avatar.properties.avgColor : null;
 	const bannerColor = banner && banner.properties.avgColor ? banner.properties.avgColor : null;
@@ -288,14 +289,14 @@ export async function updatePerson(uri: string, resolver?: Resolver | null, hint
 	logger.info(`Updating the Person: ${person.id}`);
 
 	// アバターとヘッダー画像をフェッチ
-	const [avatar, banner] = (await Promise.all<DriveFile | null>([
+	const [avatar, banner] = await Promise.all([
 		person.icon,
 		person.image
 	].map(img =>
 		img == null
 			? Promise.resolve(null)
 			: resolveImage(exist, img).catch(() => null)
-	)));
+	));
 
 	// カスタム絵文字取得
 	const emojis = await extractEmojis(person.tag || [], exist.host).catch(e => {
@@ -319,12 +320,12 @@ export async function updatePerson(uri: string, resolver?: Resolver | null, hint
 		tags,
 		isBot: object.type == 'Service',
 		isCat: (person as any).isCat === true,
-		isLocked: person.manuallyApprovesFollowers,
+		isLocked: !!person.manuallyApprovesFollowers,
 	} as Partial<User>;
 
 	if (avatar) {
 		updates.avatarId = avatar.id;
-		updates.avatarUrl = DriveFiles.getPublicUrl(avatar);
+		updates.avatarUrl = DriveFiles.getPublicUrl(avatar, true);
 		updates.avatarColor = avatar.properties.avgColor ? avatar.properties.avgColor : null;
 	}
 
