@@ -2,7 +2,7 @@ import webFinger from './webfinger';
 import config from '../config';
 import { createPerson, updatePerson } from './activitypub/models/person';
 import { remoteLogger } from './logger';
-import chalk from 'chalk';
+import * as chalk from 'chalk';
 import { User, IRemoteUser } from '../models/entities/user';
 import { Users } from '../models';
 import { toPuny } from '../misc/convert-host';
