@@ -4,6 +4,7 @@ import { detectUrlMine } from '../../../../../misc/detect-url-mine';
 import { ID } from '../../../../../misc/cafy-id';
 import { Emojis } from '../../../../../models';
 import { getConnection } from 'typeorm';
+import { ApiError } from '../../../error';
 
 export const meta = {
 	desc: {
@@ -24,6 +25,10 @@ export const meta = {
 			validator: $.str
 		},
 
+		category: {
+			validator: $.optional.str
+		},
+
 		url: {
 			validator: $.str
 		},
@@ -31,19 +36,28 @@ export const meta = {
 		aliases: {
 			validator: $.arr($.str)
 		}
+	},
+
+	errors: {
+		noSuchEmoji: {
+			message: 'No such emoji.',
+			code: 'NO_SUCH_EMOJI',
+			id: '684dec9d-a8c2-4364-9aa8-456c49cb1dc8'
+		}
 	}
 };
 
 export default define(meta, async (ps) => {
 	const emoji = await Emojis.findOne(ps.id);
 
-	if (emoji == null) throw new Error('emoji not found');
+	if (emoji == null) throw new ApiError(meta.errors.noSuchEmoji);
 
 	const type = await detectUrlMine(ps.url);
 
 	await Emojis.update(emoji.id, {
 		updatedAt: new Date(),
 		name: ps.name,
+		category: ps.category,
 		aliases: ps.aliases,
 		url: ps.url,
 		type,
