@@ -11,11 +11,11 @@ import { Users, UserProfiles } from '../../../models';
 import { ILocalUser } from '../../../models/entities/user';
 import { ensure } from '../../../prelude/ensure';
 
-function getUserToken(ctx: Koa.BaseContext) {
+function getUserToken(ctx: Koa.Context) {
 	return ((ctx.headers['cookie'] || '').match(/i=(\w+)/) || [null, null])[1];
 }
 
-function compareOrigin(ctx: Koa.BaseContext) {
+function compareOrigin(ctx: Koa.Context) {
 	function normalizeUrl(url: string) {
 		return url.endsWith('/') ? url.substr(0, url.length - 1) : url;
 	}
