@@ -25,30 +25,19 @@ export default class extends Channel {
 	@autobind
 	private async onNote(note: PackedNote) {
 		if ((note.user as PackedUser).host !== null) return;
-		if (note.visibility === 'home') return;
-		if (!this.user && note.visibility === 'users') return;
+		if (note.visibility !== 'public') return;
 
-		if (['followers', 'specified'].includes(note.visibility)) {
-			note = await Notes.pack(note.id, this.user, {
+		// リプライなら再pack
+		if (note.replyId != null) {
+			note.reply = await Notes.pack(note.replyId, this.user, {
 				detail: true
 			});
-
-			if (note.isHidden) {
-				return;
-			}
-		} else {
-			// リプライなら再pack
-			if (note.replyId != null) {
-				note.reply = await Notes.pack(note.replyId, this.user, {
-					detail: true
-				});
-			}
-			// Renoteなら再pack
-			if (note.renoteId != null) {
-				note.renote = await Notes.pack(note.renoteId, this.user, {
-					detail: true
-				});
-			}
+		}
+		// Renoteなら再pack
+		if (note.renoteId != null) {
+			note.renote = await Notes.pack(note.renoteId, this.user, {
+				detail: true
+			});
 		}
 
 		// 流れてきたNoteがミュートしているユーザーが関わるものだったら無視する
