@@ -2,7 +2,7 @@ import * as Koa from 'koa';
 import * as manifest from '../../client/assets/manifest.json';
 import { fetchMeta } from '../../misc/fetch-meta';
 
-module.exports = async (ctx: Koa.BaseContext) => {
+module.exports = async (ctx: Koa.Context) => {
 	const json = JSON.parse(JSON.stringify(manifest));
 
 	const instance = await fetchMeta(true);
