@@ -69,7 +69,8 @@ function postProcess(file: DriveFile, isExpired = false) {
 			isLink: true,
 			url: file.uri,
 			thumbnailUrl: file.uri,
-			webpublicUrl: file.uri
+			webpublicUrl: file.uri,
+			size: 0,
 		});
 	} else {
 		DriveFiles.delete(file.id);
