@@ -3,14 +3,14 @@ import renderDelete from '../../remote/activitypub/renderer/delete';
 import renderAnnounce from '../../remote/activitypub/renderer/announce';
 import renderUndo from '../../remote/activitypub/renderer/undo';
 import { renderActivity } from '../../remote/activitypub/renderer';
-import { deliver } from '../../queue';
 import renderTombstone from '../../remote/activitypub/renderer/tombstone';
 import config from '../../config';
 import { registerOrFetchInstanceDoc } from '../register-or-fetch-instance-doc';
 import { User } from '../../models/entities/user';
 import { Note } from '../../models/entities/note';
-import { Notes, Users, Followings, Instances } from '../../models';
+import { Notes, Users, Instances } from '../../models';
 import { notesChart, perUserNotesChart, instanceChart } from '../chart';
+import { deliverToFollowers } from '../../remote/activitypub/deliver-manager';
 
 /**
  * 投稿を削除します。
@@ -49,22 +49,7 @@ export default async function(user: User, note: Note, quiet = false) {
 				? renderUndo(renderAnnounce(renote.uri || `${config.url}/notes/${renote.id}`, note), user)
 				: renderDelete(renderTombstone(`${config.url}/notes/${note.id}`), user));
 
-			const queue: string[] = [];
-
-			const followers = await Followings.find({
-				followeeId: note.userId
-			});
-
-			for (const following of followers) {
-				if (Followings.isRemoteFollower(following)) {
-					const inbox = following.followerSharedInbox || following.followerInbox;
-					if (!queue.includes(inbox)) queue.push(inbox);
-				}
-			}
-
-			for (const inbox of queue) {
-				deliver(user as any, content, inbox);
-			}
+			deliverToFollowers(user, content);
 		}
 		//#endregion
 
